// P1Frontend/src/interfaces/UserInterface.ts
export interface UserInterface {
    userId?: number,
    firstName: string,
    lastName: string,
    username: string,
    email: string,
    role?: string
}

//every car needs a registered make, model, and isFourWheelDrive
//carId and userId are optional - we don't supply these values if we're inserting a car
    //carId is autogenerated. userId will be supplied by the saved logged in user data